Uudelleennimeä s3-package -> aws
