+++ server/src/main/kotlin/fi/oph/ludos/certificate/CertificateService.kt
@@ -3,8 +3,8 @@ package fi.oph.ludos.certificate
 import arrow.core.Either
 import fi.oph.ludos.Exam
 import fi.oph.ludos.auth.OppijanumerorekisteriClient
-import fi.oph.ludos.s3.Bucket
-import fi.oph.ludos.s3.S3Helper
+import fi.oph.ludos.aws.Bucket
+import fi.oph.ludos.aws.S3Helper
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
